# apps/web/public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Wedding Website API",
    "version": "1.0.0",
    "description": "API for the wedding website - manage guests, events, RSVPs, gifts, photos, and more"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/admin/templates": {
      "get": {
        "operationId": "get-admin-templates",
        "summary": "List email templates",
        "description": "Fetch all email templates from Resend",
        "tags": ["Admin - Templates"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-admin-templates",
        "summary": "Create email template",
        "description": "Create a new email template in Resend with optional variables and auto-publish",
        "tags": ["Admin - Templates"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTemplateBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/templates/{id}": {
      "get": {
        "operationId": "get-admin-templates-{id}",
        "summary": "Get email template",
        "description": "Fetch a specific email template by ID from Resend",
        "tags": ["Admin - Templates"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch-admin-templates-{id}",
        "summary": "Update email template",
        "description": "Update an existing email template's name, subject, or HTML content",
        "tags": ["Admin - Templates"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTemplateBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-admin-templates-{id}",
        "summary": "Delete email template",
        "description": "Permanently delete an email template from Resend",
        "tags": ["Admin - Templates"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/templates/{id}/publish": {
      "post": {
        "operationId": "post-admin-templates-{id}-publish",
        "summary": "Publish email template",
        "description": "Publish a draft email template in Resend, making it available for sending",
        "tags": ["Admin - Templates"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "description": "Resource UUID"
            },
            "required": true,
            "description": "Resource UUID",
            "example": "123"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/rsvp/submit": {
      "post": {
        "operationId": "post-rsvp-submit",
        "summary": "Submit RSVP",
        "description": "Submit an RSVP response for a guest party. Updates all guests with the given invite code",
        "tags": ["RSVP"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RsvpSubmitBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RsvpSubmitResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation succeeded"
          }
        },
        "required": ["success"]
      },
...
      "UpdateTemplateBody": {},
      "CreateTemplateBody": {},
...
      "RsvpSubmitBody": {
        "type": "object",
        "properties": {
          "inviteCode": {
            "type": "string",
            "description": "Invite code (required)"
          },
          "attending": {
            "type": "boolean",
            "description": "Whether the guest is attending"
          },
          "dietaryRestrictions": {
            "type": "string",
            "description": "Dietary restrictions"
          }
        },
        "required": ["inviteCode", "attending"]
      },
      "RsvpSubmitResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the RSVP was submitted"
          }
        },
        "required": ["success"]
      },
...
    },
...
  }
}